Log properly when handling requests concurrently
Previously we logged the delivery ID of the request once, and then
logged additional messages sans delivery ID, assuming they'd be
associated with the request. This doesn't work when multiple requests
are handled concurrently, in which case sequences of log messages from
different requests may be interleaved. To fix this, prefix each log
message sent from the HTTP handler with "(Delivery ID: <deliveryID>)".